new option: skipPagesStartingWithLocale
This new option 'skipPagesStartingWithLocale' allows to skip creating page clones per locale if the path already starts with a locale e.g. /en/blog/foo or /de/blog/bar

This can be the case for templates which already contain a locale coming e.g. from a .md file where people maintain the language / locale themselves (e.g. in a frontmatter block with a "lang: en" flag)

In this case those pages are already correct and can be skipped while still creating a locale-path for everything else.